+++ src/main/java/org/nrg/containers/model/server/docker/DockerServerPrefsBean.java
@@ -2,6 +2,7 @@ package org.nrg.containers.model.server.docker;
 
 import com.google.common.base.MoreObjects;
 import org.nrg.framework.configuration.ConfigPaths;
+import org.nrg.framework.utilities.OrderedProperties;
 import org.nrg.prefs.annotations.NrgPreference;
 import org.nrg.prefs.annotations.NrgPreferenceBean;
 import org.nrg.prefs.beans.AbstractPreferenceBean;
@@ -16,32 +17,36 @@ import java.util.Objects;
 
 /**
  * Docker server settings, stored as a prefs bean.
- * Deprecated in favor of storing docker server(s) in a hibernate table.
+ * Preferences ignored in favor of existing server settings in a hibernate table.
  * Use DockerServer / DockerServerService for the pojo, and
  *  DockerServerEntity / DockerServerEntityService for the entity.
- * @since 1.3
  */
 @NrgPreferenceBean(toolId = "docker-server",
     toolName = "Docker Server Prefs",
     description = "All the preferences that define a Docker Server")
-@Deprecated
 public class DockerServerPrefsBean extends AbstractPreferenceBean {
     private static final Logger _log = LoggerFactory.getLogger(DockerServerPrefsBean.class);
 
-    @Autowired
     public DockerServerPrefsBean(final NrgPreferenceService preferenceService) {
         super(preferenceService);
     }
 
-    public DockerServerPrefsBean(final NrgPreferenceService preferenceService, final ConfigPaths configFolderPaths) {
-        super(preferenceService, configFolderPaths);
+    @Autowired
+    public DockerServerPrefsBean(final NrgPreferenceService preferenceService, final ConfigPaths configFolderPaths, final
+                                 OrderedProperties initPrefs) {
+        super(preferenceService, configFolderPaths, initPrefs);
     }
 
     public void fromPojo(final DockerServerBase.DockerServer dockerServer) throws InvalidPreferenceName {
         setHost(dockerServer.host());
         setName(dockerServer.name());
         setCertPath(dockerServer.certPath());
+        setContainerUser(dockerServer.containerUser());
         setLastEventCheckTime(new Date()); // Initialize with current time
+    }
+
+    public DockerServerBase.DockerServer toPojo() {
+        return DockerServerBase.DockerServer.create(this);
     }
 
     @NrgPreference(defaultValue = "Local socket")
@@ -89,6 +94,21 @@ public class DockerServerPrefsBean extends AbstractPreferenceBean {
         }
     }
 
+    @NrgPreference(key = "containerUser")
+    public String getContainerUser() {
+        return getValue("containerUser");
+    }
+
+    public void setContainerUser(final String containerUser) {
+        if (containerUser != null) {
+            try {
+                set(containerUser, "containerUser");
+            } catch (InvalidPreferenceName e) {
+                _log.error("Error setting Docker server preference \"containerUser\".", e.getMessage());
+            }
+        }
+    }
+
     @NrgPreference
     public Date getLastEventCheckTime() {
         return getDateValue("lastEventCheckTime");
@@ -110,6 +130,7 @@ public class DockerServerPrefsBean extends AbstractPreferenceBean {
                 .add("name", getName())
                 .add("host", getHost())
                 .add("certPath", getCertPath())
+                .add("containerUser", getContainerUser())
                 .add("lastEventCheckTime", getLastEventCheckTime())
                 .toString();
     }
@@ -127,11 +148,12 @@ public class DockerServerPrefsBean extends AbstractPreferenceBean {
 
         return Objects.equals(this.getName(), that.getName()) &&
                 Objects.equals(this.getHost(), that.getHost()) &&
+                Objects.equals(this.getContainerUser(), that.getContainerUser()) &&
                 Objects.equals(this.getCertPath(), that.getCertPath());
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(getName(), getHost(), getCertPath());
+        return Objects.hash(getName(), getHost(), getCertPath(), getContainerUser());
     }
 }